mudando o tipo do Preco para double

+++ Teste Unitário/ValueObject/Precos/PrecoDeveria.cs
@@ -7,31 +7,31 @@ namespace Teste_Unitário.ValueObject.Precos
     public class PrecoDeveria
     {
         [DataTestMethod]
-        [DataRow("R$12,34", 12.34f)]
-        [DataRow("R$12,3", 12.3f)]
-        [DataRow("R$ 12,34", 12.34f)]
-        [DataRow("R$ 12,3", 12.3f)]
-        [DataRow(" R$ 12,34", 12.34f)]
-        [DataRow("R$ 12,34 ", 12.34f)]
-        [DataRow(" R$ 12,34 ", 12.34f)]
-        [DataRow("12,34", 12.34f)]
-        [DataRow(" 12,34", 12.34f)]
-        [DataRow("12,34 ", 12.34f)]
-        [DataRow(" 12,34 ", 12.34f)]
-        [DataRow("R$ 12,34,32", 1234.32f)]
-        [DataRow("12,34,32", 1234.32f)]
-        [DataRow("R$ 12!34,32", 1234.32f)]
-        public void DeveriaFormatarAStringCorretamenteNoConstrutor(string valor, float valorFormatado)
+        [DataRow("R$12,34", 12.34d)]
+        [DataRow("R$12,3", 12.3d)]
+        [DataRow("R$ 12,34", 12.34d)]
+        [DataRow("R$ 12,3", 12.3d)]
+        [DataRow(" R$ 12,34", 12.34d)]
+        [DataRow("R$ 12,34 ", 12.34d)]
+        [DataRow(" R$ 12,34 ", 12.34d)]
+        [DataRow("12,34", 12.34d)]
+        [DataRow(" 12,34", 12.34d)]
+        [DataRow("12,34 ", 12.34d)]
+        [DataRow(" 12,34 ", 12.34d)]
+        [DataRow("R$ 12,34,32", 1234.32d)]
+        [DataRow("12,34,32", 1234.32d)]
+        [DataRow("R$ 12!34,32", 1234.32d)]
+        public void DeveriaFormatarAStringCorretamenteNoConstrutor(string valor, double valorFormatado)
         {
             Preco preco = valor;
             Assert.AreEqual(valorFormatado, preco.Valor);
         }
 
         [DataTestMethod]
-        [DataRow(10f, 10f, 20f)]
-        [DataRow(10f, 30f, 40f)]
-        [DataRow(10.6f, 30.4f, 41f)]
-        public void DeveriaSomarPrecosCorretamente(float a, float b, float esperado)
+        [DataRow(10d, 10d, 20d)]
+        [DataRow(10d, 30d, 40d)]
+        [DataRow(10.6d, 30.4d, 41d)]
+        public void DeveriaSomarPrecosCorretamente(double a, double b, double esperado)
         {
             Preco precoA = a;
             Preco precoB = b;
@@ -41,10 +41,10 @@ namespace Teste_Unitário.ValueObject.Precos
         }
 
         [DataTestMethod]
-        [DataRow(10f, 10f, 0f)]
-        [DataRow(10f, 30f, -20f)]
-        [DataRow(10.6f, 30.4f, -19.8f)]
-        public void DeveriaSubtrairPrecosCorretamente(float a, float b, float esperado)
+        [DataRow(10d, 10d, 0d)]
+        [DataRow(10d, 30d, -20d)]
+        [DataRow(10.6d, 30.4d, -19.8d)]
+        public void DeveriaSubtrairPrecosCorretamente(double a, double b, double esperado)
         {
             Preco precoA = a;
             Preco precoB = b;
@@ -54,10 +54,10 @@ namespace Teste_Unitário.ValueObject.Precos
         }
 
         [DataTestMethod]
-        [DataRow(10f, 10f, 100f)]
-        [DataRow(10f, 30f, 300f)]
-        [DataRow(10.6f, 30.4f, 322.24f)]
-        public void DeveriaMultiplicarPrecosCorretamente(float a, float b, float esperado)
+        [DataRow(10d, 10d, 100d)]
+        [DataRow(10d, 30d, 300d)]
+        [DataRow(10.6d, 30.4d, 322.24d)]
+        public void DeveriaMultiplicarPrecosCorretamente(double a, double b, double esperado)
         {
             Preco precoA = a;
             Preco precoB = b;
@@ -67,10 +67,10 @@ namespace Teste_Unitário.ValueObject.Precos
         }
 
         [DataTestMethod]
-        [DataRow(10f, 10f, 1f)]
-        [DataRow(10f, 30f, 0.33f)]
-        [DataRow(10.6f, 30.4f, 0.35f)]
-        public void DeveriaDividirPrecosCorretamente(float a, float b, float esperado)
+        [DataRow(10d, 10d, 1d)]
+        [DataRow(10d, 30d, 0.33d)]
+        [DataRow(10.6d, 30.4d, 0.35d)]
+        public void DeveriaDividirPrecosCorretamente(double a, double b, double esperado)
         {
             Preco precoA = a;
             Preco precoB = b;
@@ -80,10 +80,10 @@ namespace Teste_Unitário.ValueObject.Precos
         }
 
         [DataTestMethod]
-        [DataRow(10f, 10f, 0f)]
-        [DataRow(10f, 30f, 10f)]
-        [DataRow(10.6f, 30.4f, 10.6f)]
-        public void DeveriaModularPrecosCorretamente(float a, float b, float esperado)
+        [DataRow(10d, 10d, 0d)]
+        [DataRow(10d, 30d, 10d)]
+        [DataRow(10.6d, 30.4d, 10.6d)]
+        public void DeveriaModularPrecosCorretamente(double a, double b, double esperado)
         {
             Preco precoA = a;
             Preco precoB = b;
